requirements.txt
asgiref==3.4.1
Django==3.2
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-redis==5.0.0
Pillow==8.3.2
pytz==2021.3
redis==3.5.3
sqlparse==0.4.2
